temp30/pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "knowledge-ingestion"
version = "0.1.0"
description = "Rewrite structured data with LLMs, embed via pgvector, and expose retrieval APIs."
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "fastapi>=0.110.0",
    "uvicorn[standard]>=0.27.0",
    "pandas>=2.1.0",
    "numpy>=1.24.0",
    "openpyxl>=3.1.2",
    "openai>=1.3.5",
    "requests>=2.31.0",
    "psycopg2-binary>=2.9.9",
    "pgvector>=0.2.5",
    "pydantic>=2.6.0",
    "python-dotenv>=1.0.0",
    "SQLAlchemy>=2.0.0",
    "PyMySQL>=1.1.0",
]

[project.optional-dependencies]
dev = [
    "pytest",
    "httpx",
]

[project.scripts]
knowledge-cli = "app.cli:main"
